packages/ih/pyproject.toml
[project]
name = "ih"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "cattrs>=24.1.2",
    "svcs>=24.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ih"]
only-packages = true

[dependency-groups]
dev = [
    "boto3-stubs[dynamodb,s3]>=1.35.33",
]
